feat: time_workload list as pydantic model, example in models
